Eric's comments on Roman's review: Hello Eric:
OK then; it's quite hard to satisfy all parties but since you're seeing things the way I saw them when I wrote the original text, I'm happy to revert..

So I turned the text back to what it was. Still I retained the text that demands the registering node to place an address from the prefix in full - if it owns one - in the Target Address field.

This 'trick' proved very useful in MIPv6, and could find its use here too. e.g;, it is an indication that the prefix is a directly attached stub, as opposed, e.g,,  to a remote aggregation that would be redistributed, and it allows the 6LBR to talk unicast to the registering node.

I hope that's fine with you;

Many thanks again!

Pascal

Le mar. 1 juil. 2025 à 15:35, Eric Vyncke (evyncke) <evyncke@cisco.com> a écrit :
Bonjour Pascal,

While the trick to avoid updating RFC 4861 is smart, it is also fitting a square peg into a round hole 😊

Let’s keep it simple and revert back to the previous text in section 4 of version -12 and formally updating RFC 4861.

Then resubmit and after a final check, I am approving the 2 6LO drafts (assuming that Roman clears his DISCUSS ballot of course)

Regards

-éric